Batch/Separate Arrivals

Activities for collecting a number of arrivals before they can continue processing and splitting a previously established batch.

batch(.trj, n, timeout = 0, permanent = FALSE, name = "", rule = NULL)

separate(.trj)

Arguments

.trj
the trajectory object.

n
batch size, accepts a numeric.

timeout
set an optional timer which triggers batches every timeout time units even if the batch size has not been fulfilled, accepts a numeric or a callable object (a function) which must return a numeric (0 = disabled).

permanent
if TRUE, batches cannot be split.

name
optional string. Unnamed batches from different batch activities are independent. However, if you want to feed arrivals from different trajectories into a same batch, you need to specify a common name across all your batch activities.

rule
an optional callable object (a function) which will be applied to every arrival to determine whether it should be included into the batch, thus

Value

Returns the trajectory object.