Select Resources

Activity for selecting a resource for a subsequent seize/release or setting its parameters (capacity or queue size).

select(.trj, resources, policy = c("shortest-queue", "round-robin",
  "first-available", "random"), id = 0)

Arguments

.trj
the trajectory object.

resources
one or more resource names, or a callable object (a function) which must return one or more resource names.

policy
if resources is a character vector, this parameter determines the criteria for selecting a resource among the set of policies available: 'shortest-queue' selects the least busy resource, 'round-robin' selects the resources in order cyclically, 'first-available' selects the first resource available, and 'random' selects one at random.

id
selection identifier for nested usage.

Value

Returns the trajectory object.

See also

seize_selected, release_selected, set_capacity_selected, set_queue_size_selected.